Capture analysis states in classes for greater modularity. (#1)
This is intended to improve isolation of analysis states, in particular
when an on-file analysis is reloaded over an existing state in memory.
The use of an explicit state also supports multiple analyses at once.

Initialization of the visualization workers is now the responsibility of 
the constructors for the individual UMAP/t-SNE state classes. This
also improves isolation but requires some more effort to track the 
initialized workers so that we can shut them down properly.

Each state also explicitly has a method to free the Wasm heap,
which should allow for better memory management in long-lived instances.