Reorganized code into subdirectories for more independent imports.
We want to minimize imports in directories outside of the current one,
i.e., minimize the number of "../" uses in the import statements. This
makes it easier to track which files depend on other files.

The major exception is when a file needs to pull in something from abstract/,
which belongs in its own directory for easier management by build.sh.
A few exceptions are also noted for importing stuff from readers/.