Use valkana to validate all kana state files in tests.
This should improve the chance of producing valid output files in
bakana-based applications. Note that some of the legacy kana files were
never valid so they do not pass the new checks; as such, we have simply
omitted those tests for older versions of the format.

The format version number is now exported, primarily for use in tests
but also to allow users to check the version of the generated files.