Refactored how InputsState processes its parameters.
The matrix loading, blocking factor generation and subsetting are now
implemented as separate steps, each with their own checks for whether the
relevant parameters have changed. This allows us to, e.g., change the
sample_factor without requiring a full re-read of the data files. A side-effect
of this refactoring is that we need to cache a copy of the full matrix, so
that we can freely re-run the subsetting step.

Speaking of which, specification of subset indices has been moved out of the
compute() arguments and into its own method. This avoids the need to check for
changes in (potentially long) arrays when running compute(). It also represents
a philosophical difference between the parameters in the compute() arguments,
which are more-or-less agnostic to the dataset; and the indices, which are
specific to a dataset and require special treatment.  This change aligns
'inputs' with how 'custom_selections' handles its index arrays.