Report summaries of per-cell annotations in the preflight response. (#24)
This allows GUIs to present a decent choice of subsetting options prior to
actually running the analysis. For efficiency, we also switch to new scran.js
methods for extracting MatrixMarket data - in this case, we only have to use
the dimensions, so we make sure to only parse the headers.