Updates to the readers for the new ScranMatrix output. (#30)
The new initialize methods from scran.js will now report the row identities
separately from the ScranMatrix, so all reader functions need to be modified to
accept an additional 'row_ids' argument. The isReorganized() and identities()
methods are deprecated, so all readers need to avoid using them. Serialization
and loading of the InputsState also needs to be updated for separate row IDs.