Enable reading from RDS files containing SummarizedExperiments. (#31)
Pretty much as it says on the tin. This involved adding a separate reader that
peels the assays, rowData and colData out of the usual SE structure, with
support for GRanges and GRLs in the rowRanges. Updated the downloads and tests
to obtain and use new RDS files for some of the datasets. Note that currently
alternative experiments are not supported.