Refactor the file format readers to simplify the contract. (#33)
All dataset readers are now instances of ES6 classes. Users are now expected to
construct these class instances before feeding them into the InputsState
constructor (and runAnalysis). The switch to classes is motivated by the
ability to cache objects for each instance of each dataset type (e.g., file
handles, pre-built structures) for quick re-use across loading and preflight.

In a related move, we export the SimpleFile class to give users a mechanism for
representing files (from paths, Uint8Arrays, or File objects). SimpleFiles can
be used as inputs for constructing the new Dataset classes without worrying
about the environment, i.e., browser vs Node.js. We support a variety of
representations for different purposes: the Uint8Array input is the most
general but is memory-intensive compared to paths or Files. The SimpleFile
replaces the old LoadedFile, the code for which has been removed.

As for the dataset contract itself: the per-cell and per-feature annotations
are now expected to be returned in bioconductor's DataFrame class. This makes
it easier to report the number of features/cells when there are no annotations,
which would have previously yielded an uninformative empty object. All code
and tests have been updated so that they work with DataFrame columns.

We now use an empty string for the default modality. The modality guessing
(i.e., RNA or ADT) is now pushed out of the readers and up to InputsState.
Dataset readers should now faithfully report the modality name and bakana will
try to figure it out, which provides a central point for these guesses.

We now rely on scran.js to remove files from the virtual/native file systems.

We now replace all instances of ArrayBuffers with Uint8Arrays for file
representation. This is to be generally consistent across the codebase and to
avoid having to guess whether an ArrayBuffer can be used in certain places.