Shift responsibility for CORS proxy setup to the application.
A proxy isn't even needed for Node.js, so it doesn't make sense that
it's hard-coded into the downloader for the cell labelling step.

We change the contract for the downloader to allow any return value that
can be used in the SimpleFile constructor. This improves consistency
but means that ArrayBuffers need to switch to Uint8Arrays.