Stream (gzip-compressed) lines and tables for improved memory efficiency. (#37)
This introduces the new readTable2 and readLines2 functions, which use
streaming to avoid the need to (i) storing the entire decompressed file in
memory, and to (ii) convert the decompressed buffer to a giant string. The
latter is especially relevant as the V8 engine breaks with very large strings.

For Node.js, we additionally support streaming of the file's contents from
disk. This further improves memory efficiency by avoiding any storage of the
compressed file as an in-memory buffer.

All internal calls to readTable and readLines have been replaced by their
v2 equivalents. This necessitates an async-ifying of some functions as the
streams operate via callbacks. On the plus side, it means that we finally
have some checks for interoperability with async Dataset.load/annotations.

Closes #36.