Eliminate the summary() method in favor of step-specific result getters. (#41)
The idea is to empower applications to write their own callbacks after each
step rather than to be restricted to the result formats in summary().

On that note, we also updated the tests to handle the loss of summary(). This
involved implementing a more rigorous set of checks for sane results from each
step, as well as a better method to compare reloaded and original state
objects. These tests are now applied consistently throughout the suite.

Also caught and fixed a few bugs after running the more rigorous tests:

- Correct indices for the versus mode with custom selections.
- Don't forget to perform blocking for custom selection markers.
- Proper unserialization of corrected PCs for older kana file formats.
- Propagate row_ids when creating a directly-subsetted InputsState.
- Getter methods for mimics now have sensible destructured defaults.