Wait for the t-sne, umap workers to be ready during load.
This (probably) fixes a few test errors where we start sending messages to the
workers before they are ready to receive, causing jest to hang.

Also shifted the animation tests to a separate function so that it is called
more sparingly - no need to call it in all the basic result checks.