Shift responsibility for modality definition back to readers.  (#42)
This avoids the need for guessing within InputsState by requiring the readers
to emit modalities within a controlled vocabulary (i.e., "RNA" and "ADT"). In
fact, even the readers no longer need to guess, as the choice of RNA and ADT
modality is specified by the user within the DataSet instance. This eliminates
one potential source of ambiguity by making it explicit at the user level.

Similarly, each Dataset is now also responsible for specification of the
primary feature identifier. This is relevant for picking the feature annotation
to use when combining multiple datasets by the intersection of the feature
spaces. SE and H5AD readers can additionally specify the assay(s) containing
the count matrix.  Again, this eliminates a potential source of ambiguity.

All of these options are encoded as options in the constructor, modifiable in
new setter methods on each Dataset instance. Options are now returned in the
abbreviate() and serialize() methods; they are saved to the kana files; and 
they are restored and passed to the constructor during loadAnalysis().

validateAnnotations() has been removed as the preflight is application- and
dataset-dependent. Instead, each Dataset instance now has a summary() method
(replacing the annotations() method) that provides more flexibility in the
return value. SE readers can return assay names for each modality, while
10X readers can just return straight DataFrames; the application is now 
responsible for deciding what to do with this.

Finally, we just return the cached feature/cell annotations directly in
summary/load, rather than making an unnecessary copy, under the assumption that
bakana will not mutate them.