Replace mimics with manually filled Result objects. (#43)
This should - to some extent - reduce the discrepancies between direct and
reloaded states. It's not quite feature-complete as the kana files don't store
all the information required for faithful recovery, but it does at least mean
that we don't have to worry whether the data is on the Wasm heap or not.