Support analyses with CRISPR screening as a separate modality. (#47)
This involves adding dedicated states for quality control, normalization
and PCA, mirroring what is done for ADTs. There are now options for
cell filtering to use the CRISPR QC if available (on by default) as well
as to use the CRISPR PCs in the combined embddings (off by default).

The combined_embeddings parameters have been simplified so that the
weights for each modality are separate arguments into compute(). This 
avoids the need to check for null-ness and protect against pass-by-ref
behavior in the parameters.

Also bumped out the timeout to account for the extra tests.