More fixes to worker messaging to try to avoid hanging unit tests.
The main change involves handling the case where the worker's callback fires
before the resolve/reject set-up code is evaluated inside the sendTask Promise.
In such cases, there would be no resolve/reject functions to call, and the
Promise would just be awaited forever. We avoid this by making sure that the
order of the setup versus the worker response doesn't matter. 

An accompanying change is to add a dead/not-dead flag to each worker, so that
they explicitly throw an error if they are passed a message after they are
killed. This probably doesn't have much to do with the unit testing stalls, but