Further updates for kana format v3.
Also migrated updateRowIdentities over from scran.js. Changed its behavior to
return a new vector rather than permuting in-place, just in case the new vector
needs to be of different length. Missing values are also filled in as
placeholders if there are new features that aren't present in the old results.