Re-run all steps when runAnalysis'ing a loaded analysis.
This avoids headaches with mixing of results from different versions, where
only a few steps might be re-run if only some of the parameters are changed.

From a developer perspective, this also mitigates problems with backwards
compatibility where older files might not contain the results needed in later
versions of the pipeline. I suppose we could try to fill in these results,
but that requires us to diff the analysis steps, which would be a pain.

Of course, applications using the reloaded state will still have to deal with
missing results, but they can just show, e.g., an "unavailable" error message,
which is probably the cleanest solution here.

Also closes #45.