Support calculation of feature set enrichment statistics. (#51)
This is implemented as a new FeatureSetEnrichmentState that supports on-demand
calculation of enrichment statistics and feature set scores. We use an on-demand
approach, rather than computing everything upfront, because the enrichment tests
depend on the user's choice for the marker rankings. Similarly, the feature set
scores are dense and not cheaply computed for all sets.

As befits an on-demand model, we don't bother serializing and loading anything
but the parameters. Everything can be recomputed cheaply from the saved marker
statistics, and I don't want to add more complexity to the kana file contract.