Dump state's results into an ArtifactDB-compatible SCE for external use. (#53)
This provides a more "analysis-ready" representation for further inspection,
rather than having to parse out the components of the kana file. It is also
more easily compatible with ArtifactDB-related systems for easier sharing.

This change feeds into the longer-term picture described in #52, whereby the
new representation can be used in read-only applications to visualize and
explore the results. This relieves us of the burden of storing excessive
content to faithfully replicate the bakana state (with back-compatibility).

Note that, while we have added a test script for the validity of the saved
files in 'miscellaneous/reload.R', we don't actually run this in the CI. It
takes too long to get through the JS tests, and then upload 200 MB's worth of
artifacts, and then download those artifacts, pull the alabaster image, etc.