Replace saveAnalysis with serializeConfiguration. (#54)
This represents the new strategy (discussed in #52) whereby we only save the
analysis configuration, i.e., parameters and datasets. Recovery of the analysis
state during unserialization involves re-running the analysis from its config,
which grants some future-proofing to changes in the format of the results. 

In particular, we no longer include the analysis results in the unserialization
process, as it is too difficult to constrain pipeline updates so that they
remain compatible with old versions of the result files. Instead, results are
now the responsibility of saveSingleCellExperiment.

Greatly streamlined the tests, updated the README.