All compute() steps now accept a parameters object.
This greatly simplifies the passing of arguments in runAnalysis; it also
gives each compute() method the opportunity to gracefully handle changes
to the argument names/number when unserializing old configurations.
Otherwise, any back-compatibility hacks would have to be implemented in
runAnalysis or earlier, which spreads the logic around too much.