Introduce the Result reader concept to examine pre-analyzed datasets. (#55)
This allows kana to re-use a lot of the same code for the explore-only mode;
specifically, to pull out existing dimensionality reduction results from H5AD
or RDS files. The Results themselves don't have anything to do with the rest of
the bakana codebase, but we might as well implement and test them here rather
than relying on some kind of testing within kana itself.