Support factor encodings in more recent H5AD files.
Also return all columns in the row annotations, rather than prefiltering it to
the IDs and symbols; the caller can decide what to do with the extra columns.
This allows us to consolidate the code for loading DFs (+ factors) from file.

We also set '_index' as the default choice for the primary feature ID column.