More explicit parametrization of the matching to mitochondrial genes. (#57)
This gives the caller more control over which annotation columns are actually
used for matching to mitochondrial genes for RNA QC (or to isotype controls,
for ADT QC) and what type of features should be used , e.g., Ensembl, symbols.
For convenience, the default behavior is to still figure out the best column
automatically, as it is painful to ask the caller to have this knowledge
ahead of time. However, this automation can be turned off if the caller knows
better... which can also improve efficiency slightly.

On that note, the mitochondrial gene lists are no longer baked into this
package (or in scran.js). Instead, they are pulled from a dedicated external
repository. This should reduce the bundle size, especially as the number of
organisms increases; at the cost of requiring more network traffic, but we
do that anyway for the feature sets and singlepp references, so whatever.