Refactored handling of the feature ID parameters for gene set testing. (#58)
In particular, we add an automatic= option to guess the column and feature
type. We also add a species= option to restrict the collections to the
relevant ones for the species in the dataset, so as to avoid reporting 
enrichment results for irrelevant collections. This brings us into line
with the mitochondrial feature parametrization for rna_quality_control.

Also, all the getters are now standard functions, which simplifies their use.
The async'ness associated with the preparation of the collections is now fully
concentrated within compute(); this is a consequence of also centralizing the
automatic choice of various parameters. We also cache the prepared collections
to avoid any possible need to re-prepare the collections in the getters.