Standardize feature parameter handling in the cell labelling step. (#59)
We add an automatic= option to indicate that the feature column and type should
be automatically inferred from the rowData. But if required, the caller can set
this false to allow them to explicitly choose these parameters for greater
control. We also consolidate the species-specific references into a single
argument, with a new species= option to filter out irrelevant references.

We also converted compute() to an async function to avoid long promise chains.
This allows us to turn fetchResults() back into a normal method, on the premise
that the caller has already await'ed for compute() to finish. We also added a
method to fetch the number of shared features for each reference, to allow
callers to gauge the reliability of the predicted labels.

We added a dedicated test suite to check that the cell labelling responds
correctly to the various combinations of parameters. To save time, we no longer
perform cell labelling in the other runAnalysis() calls, as the dedicated
tests should provide enough coverage.

Also updated the fetch URL for the references, and added Entrez support.