Clone some parameters to avoid pass-by-reference links.
This applies to all situations where the parameters are stored inside a
class instance. Any parameter that might be a Javascript reference
object should be CLONE'd before being stored inside the class, and
similarly CLONE'd on output via fetchParameters.  This ensures that
users cannot accidentally mutate parameters inside the class due to
pass-by-reference behavior from an external reference to that object.