Support arbitrary input vectors in the standalone classes. (#63)
General arrays and TypedArrays can now be passed to the Standalone
constructors, e.g., as groups or blocking factors. Factorization is handled
internally, rather than requiring the caller to do it, which should make them a
bit easier to use. Invalid values are also detected and transparently filtered
out (along with appropriate subsetting of the associated matrix data).