Avoid tight coupling between feature_set_enrichment and other states.
feature_set_enrichment doesn't actually depend on marker_detection, nor
does custom_selections depend on feature_set_enrichment, in the sense
that none of these participate in the other's compute() step. So, the
current arrangement (storing it as an upstream dep) is not appropriate.

It seems fairly trivial for users to pass the output of the marker
detection or custom selection's fetchResults() directly to
feature_set_enrichment's computeEnrichment() step. This avoids coupling,
is understandable, and doesn't require learning more methods.

As a result, we remove the compute-from-markers methods from the
feature_set_enrichment step. We also remove enrichment-based methods
from the custom_selections step. This means that we don't need to rename
the methods for extracting regular/versus-mode markers, because there's
no need to disambiguate between marker and enrichment results anymore.