Switch to gesel's database for the feature set enrichment step. (#65)
In the short term, this gives us access to the MSigDB collections in gesel. In
the long term, this means that we only have to maintain one gene set database,
rather than having separate ones for gesel and kana.

This is a breaking change as all sets for all collections are now reported
together, rather than sets being stratified by collection. This is probably a
better choice anyway, as users can more easily rank genes between collections.