Overhauled how parameters are passed into the dataset/result readers. (#68)
Rather than having one setter per option, callers can now pass an object to the
new setOptions() method, which makes life a little easier. 

We also removed the options= object from most constructors to avoid the
implication that those options can affect the output of summary(); this is
because summary()'s output should be constant throughout the lifetime of an
instance, in order to feed the setters to affect load().

While we're here, we also added a defaults() static method and an options()
method to all reader classes. This allows callers to inspect the defaults
and to check what parameters are already set before calling setOptions().