Separated per-gene results from SCE in the dump.
This should make life a little easier for loading in clients, as we
don't have to load all the marker tables at the same time.

We now dump the existing column annotations for posterity. We
add prefixes in front of kana-computed fields to better distinguish between those
and the existing fields.

To simplify matters, any blocking factor is always reported by
converting the factor back to an array of its original levels.
This avoids arguing about indices and such.

Custom selections now have a reportOneIndex option to specify whether
they should be reported as 1-based indices, e.g., for use in R.
We also report them as a boolean filter in the column data, for easier
use in subsetting.