Overhauled the layout of the saved ArtifactDB ZIP file.
The biggest change is that we separated per-gene results from the SCE.  This
should make life a little easier for loading in clients, as they don't have to
load all the marker tables at the same time as the SCE. Rather, they can just
load the markers for clusters of interest whenever they want to look at them.

We now include the existing column annotations in the SCE's colData for nicer
downstream usability. As a consequence, we of add 'kana::' prefixes in front of
kana-computed fields to distinguish between those and existing fields.  To
simplify matters, any blocking factor is always reported by converting the
factor back to an array of its original levels. 

Custom selections now have a reportOneIndex option to specify whether they
should be reported as 1-based indices, e.g., for use in R.  We also report them
as a boolean filter in the column data for more general use.