Dump modality-specific column stats in the main exp's colData.
This is necessary to ensure that all of the computed statistics from the
analysis mode are actually readable by bakana itself in explore mode,
given that the column data of the alternative exps is currently ignored.

That said, it is cleaner to have these statistics in the alternative exp
alongside the same modality's assay data, so we provide the option to
saveSCE to dump it there. But that's not the default behavior anymore.