Refactored CellLabellingState for more general re-usability. (#74)
This now uses the same approach as the FeatureSetEnrichmentState, where the
compute() method just sets up the references and the caller needs to use a
separate method to actually compute (non-cached) results for their inputs. This
allows a single State object to accept inputs from a variety of sources, e.g.,
marker results for clusters or custom selections. Alternatively, we can
aggregate on user-specified cell annotations to label arbitrary groups.

We added a new CellLabellingStandalone class that uses the internal functions 
that are now split off from the refactored CellLabellingState. Like its
equivalents (e.g., FeatureSetEnrichmentStandalone), this allows callers to
perform cell type labelling outside of the runAnalysis() framework.

A consequence of these changes is that the cell labelling results are not
guaranteed to be generated a a result of runAnalysis(), so we removed
unnecessary dumping of the labelling results in the SCE. Perhaps at some point
we may add them back in as separate DataFrames a la marker_detection.

This refactor involved pulling in the latest scran.js, which has a few relevant
changes for cell labelling. Specifically, the output object now reports the
scores associated with each label for improved diagnostics. This requires
callers of the new computeLabels() method to free() the output once done.