Don't report row IDs from dataset readers or use them in InputsState. (#75)
The concept of row identifiers (and associated permutations) should be strictly
internal to the dataset readers, and shouldn't be used anywhere in the main
bakana pipeline, including in InputsState. This involves a bit of work to
refactor the test suite as the checks for correct permutations were based on
the IDs; these tests are now based on mapping the feature names.