Switch to the alabaster format for saving analysis results. (#83)
This moves away from the old ArtifactDB formats and to the new alabaster
formats, allowing us to use kana's outputs with the latest alabaster/dolomite
readers and writers in R/Python, respectively. We cross-check our alabaster
writer with validation code in R to make sure that we're compatible.

This update requires the latest scran.js, which requires a litany of similar
updates throughout the pipeline. User-visible changes include:

- Discard vectors for QC are now keep vectors.
- Method names have changed for QC and marker results.
- k-means cluster initialization now uses variance partitioning.

Saving/writing to alabaster is handled by the new jaspagate dependency.