Streamlined tests for the various dataset/result readers. (#84)
We define a test suite to apply to each class that purports to implement the
custom Dataset or Result interfaces. This consolidates the test logic and means
that we only need to test the immediate outputs for each class, instead of
running the entire analysis pipeline for each Dataset class.

We also fixed a bug in the processing of integer *Experiment options, where
they were not being mapped to their names in some previewPrimaryIds() methods.