Improved handling of main experiment name in SE, alabaster readers.
The main experiment now takes precedence over any alternative experiments with
the same name in objects returned by summary() and load(). This avoids
clobbering of the main experiment, given that it is probably more important
than any alternative experiment. (The Dataset and Result contracts require some
kind of modality name, so if multiple main/alternative experiments have the
same name, we unfortunately must make a decision on which one to report.)

Additionally, the SE reader now retrieves the main experiment name from the
supplied SCE object rather than defaulting to an empty string. This allows us
to simplify the docs by removing the confusing special case where an empty
string refers to the main experiment instead of a emptily-named altExp.