All State classes now have a defaults() method.
The defaults are used to fill in any parameters that are not explicitly
supplied as fields of parameters= in compute(). We also throw an error if
unknown fields are present, to avoid bugs due to misnamed parameters.

We update the analysisDefaults() function with calls to defaults() for all
steps. We also use each step's 'step_name' to centralize the definitions.